Implemented onLoad() in a bunch of windows/popups reset them on exit

+++ src/main/java/com/farkasch/barista/gui/codinginterface/SwitchMenu.java
@@ -45,7 +45,9 @@ public class SwitchMenu extends HBox {
     }
     currentlyActive = widget;
     persistenceService.setActiveFile(currentlyActive.getFile());
-    persistenceService.refreshSideMenu();
+    if(persistenceService.getOpenProject() == null){
+      persistenceService.refreshSideMenu();
+    }
   }
 
   public void switchToFile(File file){
@@ -62,9 +64,11 @@ public class SwitchMenu extends HBox {
   }
 
   public void removeFile(int index) {
-    persistenceService.removeOpenFile(((SwitchMenuItem)(getChildren().get(index))).getFile());
-    persistenceService.addRecentlyClosed(((SwitchMenuItem)(getChildren().get(index))).getFile());
-    persistenceService.refreshSideMenu();
+    if(persistenceService.getOpenProject() == null){
+      persistenceService.removeOpenFile(((SwitchMenuItem)(getChildren().get(index))).getFile());
+      persistenceService.addRecentlyClosed(((SwitchMenuItem)(getChildren().get(index))).getFile());
+      persistenceService.refreshSideMenu();
+    }
     getChildren().remove(index);
   }
